Update patch set 1
Patch Set 1:

A few questions ...
Do we need coverage limits on the sub-projects?  Or if we put a single coverage limit on the top-level pom will that include code fro m the subprojects?

I'm a bit concerned that you've set some of the sub-project limits to 100%.  That doesn't give us any wiggle room if something is added that isn't easily unit testable.  

Wondering if it would make sense to set these to like 90%?  Or possibly just set limits on the top-level pom to whatever we are at now?

Patch-set: 1
Label: Verified=0